+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.57
 - `Ctrl + Click` in edit mode now works on the whole portion of links like `[QOwnNotes - cross-platform open source notepad](http://www.qownnotes.org/)`, not just on the URL alone
 - added a menu entry to the new knowledge base in the help menu
+- you can now change the font of the note text edit and the note markdown view in the settings (for [Issue #19](https://github.com/pbek/QOwnNotes/issues/19))
 
 ## 0.56
 - added a shortcut list to the help menu and to the startup notes